Hacked android-simple to add a very large number of classes.
This is used for timing purposes and to also make it easier to scan
through the Dalvik logs to look for problematic class loading
behaviour.

The way this is structured is that there is a new HomeActivity2 class
which requests 500 autogenerated dummy classes.  HomeActivity (the
original) is still the main entry point into the app and HomeActivity2
is only loaded when clicking a button in HomeActivity.  I then
produced logging statements that indicate the transition and also log
various stages of the Dagger injection process to correlate with
Dalvik's LOG_CLASS_LOADING feature on a custom AOSP build.